Create Wi-Fi Help Commands case script
First draft to see if everything is going to go correctly. Turns out that there were a few mistakes that will be corrected.

Goal is to make a case script to where I use this instead of having to go to the man page. Just need the steps that I know I will need or something that will help me get there if I do not know what to look for or what command to type in for man.